# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Oddiya Travel Planning API
  description: |
    AI-powered travel planning and video generation service API.
    
    ## Features
    - OAuth2 authentication (Google, Apple)
    - AI-powered travel plan generation using AWS Bedrock
    - Photo to video shorts generation
    - Place discovery and recommendations
    - Real-time notifications
    
    ## Authentication
    Most endpoints require Bearer token authentication.
    Include the token in the Authorization header: `Bearer {accessToken}`
  version: 1.0.0
  contact:
    name: Oddiya API Support
    email: api-support@oddiya.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: https://api.oddiya.com/api/v1
    description: Production server
  - url: https://api-test.oddiya.com/api/v1
    description: Test server
  - url: http://localhost:8080/api/v1
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /travel-plans:
    get:
      tags:
        - Travel Plans
      summary: List travel plans
      description: Get list of user's travel plans
      operationId: getTravelPlans
      parameters:
        - $ref: '#/components/parameters/StatusParam'
        - $ref: '#/components/parameters/StartDateParam'
        - $ref: '#/components/parameters/EndDateParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SortParam'
        - $ref: '#/components/parameters/OrderParam'
      responses:
        '200':
          description: Travel plans retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TravelPlanList'
        '401':
          $ref: '#/components/responses/Unauthorized'
    
    post:
      tags:
        - Travel Plans
      summary: Create travel plan
      description: Create new travel plan manually
      operationId: createTravelPlan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePlanRequest'
      responses:
        '201':
          description: Travel plan created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TravelPlan'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /travel-plans/{planId}:
    get:
      tags:
        - Travel Plans
      summary: Get travel plan details
      description: Get detailed information of a specific travel plan
      operationId: getTravelPlanById
      parameters:
        - $ref: '#/components/parameters/PlanIdParam'
      responses:
        '200':
          description: Travel plan details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TravelPlanDetailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    
    put:
      tags:
        - Travel Plans
      summary: Update travel plan
      description: Update existing travel plan
      operationId: updateTravelPlan
      parameters:
        - $ref: '#/components/parameters/PlanIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePlanRequest'
      responses:
        '200':
          description: Travel plan updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TravelPlanDetailed'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    
    delete:
      tags:
        - Travel Plans
      summary: Delete travel plan
      description: Delete existing travel plan
      operationId: deleteTravelPlan
      parameters:
        - $ref: '#/components/parameters/PlanIdParam'
      responses:
        '204':
          description: Travel plan deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  parameters:
    PlanIdParam:
      name: planId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Travel plan ID
...
    StatusParam:
      name: status
      in: query
      schema:
        type: string
        enum: [draft, confirmed, completed, cancelled]
      description: Plan status filter
    StartDateParam:
      name: startDate
      in: query
      schema:
        type: string
        format: date
      description: Start date filter
    EndDateParam:
      name: endDate
      in: query
      schema:
        type: string
        format: date
      description: End date filter
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Results per page
    SortParam:
      name: sort
      in: query
      schema:
        type: string
        enum: [createdAt, startDate]
        default: createdAt
      description: Sort field
    OrderParam:
      name: order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: desc
      description: Sort order
  schemas:
...
    CreatePlanRequest:
      type: object
      required:
        - title
        - destination
        - startDate
        - endDate
      properties:
        title:
          type: string
          maxLength: 200
        destination:
          type: string
          maxLength: 100
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        budget:
          type: number
          minimum: 0
        companions:
          $ref: '#/components/schemas/Companions'
        description:
          type: string
          maxLength: 1000
    UpdatePlanRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 200
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        budget:
          type: number
          minimum: 0
        status:
          type: string
          enum: [draft, confirmed, completed, cancelled]
...
    ApiResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
        data:
          type: object
          nullable: true
        meta:
          $ref: '#/components/schemas/ResponseMeta'
        error:
          $ref: '#/components/schemas/ErrorDetail'
          nullable: true
    ErrorDetail:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: object
        timestamp:
          type: string
          format: date-time
    ResponseMeta:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        version:
          type: string
...
    TravelPlanList:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                items:
                  type: array
                  items:
                    $ref: '#/components/schemas/TravelPlanSummary'
                pagination:
                  $ref: '#/components/schemas/Pagination'
...
    TravelPlan:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        destination:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        budget:
          type: number
        status:
          type: string
          enum: [draft, confirmed, completed, cancelled]
        createdAt:
          type: string
          format: date-time
    TravelPlanSummary:
      allOf:
        - $ref: '#/components/schemas/TravelPlan'
        - type: object
          properties:
            thumbnailUrl:
              type: string
              format: uri
            participantCount:
              type: integer
            completionRate:
              type: number
              minimum: 0
              maximum: 1
    TravelPlanDetailed:
      allOf:
        - $ref: '#/components/schemas/TravelPlan'
        - type: object
          properties:
            companions:
              $ref: '#/components/schemas/Companions'
            dailyPlans:
              type: array
              items:
                $ref: '#/components/schemas/DailyPlan'
            participants:
              type: array
              items:
                $ref: '#/components/schemas/Participant'
            statistics:
              $ref: '#/components/schemas/PlanStatistics'
            metadata:
              $ref: '#/components/schemas/PlanMetadata'
            updatedAt:
              type: string
              format: date-time
    Companions:
      type: object
      properties:
        adults:
          type: integer
          minimum: 1
          default: 1
        children:
          type: integer
          minimum: 0
          default: 0
...
    DailyPlan:
      type: object
      properties:
        day:
          type: integer
        date:
          type: string
          format: date
        title:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/ItineraryItem'
        estimatedCost:
          type: number
        totalDistance:
          type: number
          description: Total distance in km
    ItineraryItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        sequence:
          type: integer
        time:
          type: string
        placeName:
          type: string
        placeId:
          type: string
        category:
          type: string
        duration:
          type: integer
          description: Duration in minutes
        cost:
          type: number
        description:
          type: string
        location:
          $ref: '#/components/schemas/Location'
        tips:
          type: string
    Location:
      type: object
      properties:
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        address:
          type: string
...
    Participant:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        nickname:
          type: string
        role:
          type: string
          enum: [owner, participant]
        joinedAt:
          type: string
          format: date-time
    PlanStatistics:
      type: object
      properties:
        totalPlaces:
          type: integer
        totalDistance:
          type: number
        estimatedCost:
          type: number
    PlanMetadata:
      type: object
      properties:
        generatedBy:
          type: string
          enum: [ai, manual]
        modelVersion:
          type: string
        confidence:
          type: number
          minimum: 0
          maximum: 1
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
  responses:
    BadRequest:
      description: Bad request - Invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            success: false
            data: null
            error:
              code: "ERR_VALIDATION"
              message: "Validation failed"
              details:
                - field: "startDate"
                  message: "Start date must be in the future"
    Unauthorized:
      description: Unauthorized - Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            success: false
            data: null
            error:
              code: "ERR_UNAUTHORIZED"
              message: "Authentication required"
...
    NotFound:
      description: Not found - Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            success: false
            data: null
            error:
              code: "ERR_NOT_FOUND"
              message: "Resource not found"
...